Controllers: rename required-authed-user -> auth-utils, ref imports
