+++ app/controllers/stories-controller/story-controller/story-claps-controller/index.js
@@ -2,7 +2,7 @@ const express = require('express');
 
 const { injectStoryClap } = require('./story-claps-middleware');
 const { ReaderClapController } = require('./reader-clap-controller');
-const { requireAuthedUser } = require('../../../require-authed-user');
+const { requireAuthedUser } = require('../../../auth-utils');
 const { exchangeSlugForUser } = require('../../../users-controller/user-controller');
 const { clappedReadersHandler, clapForStoryHandler } = require('./story-claps-route-handlers');
